refactor: phase 9 — cleanup, docs, and coverage test
- Move core/git_health.py → domains/warehouse/git_health.py (git health
  checks are warehouse-specific domain logic, not cross-domain primitives)
- Update all call-sites in cli/, domains/, and tests to use new import path
- Verify utils/ contains only git.py, display.py, interaction.py (145 lines
  total, well under 500 target)
- Verify core/ contains exactly manifest/, settings.py, exceptions.py,
  gitignore.py
- Update AGENTS.md: replace 'CLI Layer Discipline' section with 'Domain
  Layer' section pointing to layered-architecture spec; update Adding a New
  CLI Command brief with correct domain paths
- Update knowledge/facts/repository-structure.md with final four-layer tree
- Update knowledge/decisions/follow-global-python-standards.md: fix stale
  import example (beacon.warehouse → beacon.domains.warehouse); add
  domain-layer import discipline rules with spec reference
- Add test_read_new_template_falls_back_to_templates_dir to cover the
  TEMPLATES_DIR disk-path branch in WarehouseUpgrader._read_new_template
  (all 22 prior upgrader tests use template_overrides, bypassing disk lookup)
- File follow-up issue #92 for optional import-linter rule (Decision 6
  mechanism 3)
